Allow for spaces in/around the command prefix
Currently, the `CommandHandler#findCommand` method retrieves the command name with this line:
```kotlin
val s = message.split(" ")[0]
```

However, this prevents using spaces in the command prefix, or allowing for an optional space after the prefix. I believe the following code change will work better, as it allows for both of the above.